chore: profilecard mediaquery, loading status
