+++ src/app/myactivities/myactivities.module.css
@@ -37,3 +37,13 @@ line-height: 42px;
   gap: 24px;
   margin-top: 142px;
 }
+
+.profileCard {
+  display: flex;
+}
+
+@media (max-width: 430px) {
+  .profileCard {
+    display: none;
+  }
+}
