+++ src/app/myactivities/page.tsx
@@ -2,16 +2,15 @@
 
 import ProfileCard from '@/components/ProfileCard/ProfileCard';
 import MyActivityContainer from './components/myactivitycontainer';
-import styles from './myactivities.module.css'
+import styles from './myactivities.module.css';
 
-export default function MyActivities(){
-    
-
-    return(
-        <div className={styles.page_container}>
-            <ProfileCard activeTab='myactivities' />
-            <MyActivityContainer />
-        </div>
-    )
+export default function MyActivities() {
+  return (
+    <div className={styles.page_container}>
+      <div className={styles.profileCard}>
+        <ProfileCard activeTab='myactivities' />
+      </div>
+      <MyActivityContainer />
+    </div>
+  );
 }
-
