+++ src/app/mynotification/components/Calendar.module.css
@@ -43,7 +43,6 @@
 .reservationWrapper {
   display: flex;
   flex-direction: column;
-  gap: 4px;
   width: 100%;
 }
