+++ src/app/mynotification/components/ReservationInfoByStatus.module.css
@@ -2,6 +2,7 @@
   display: flex;
   flex-direction: column;
   padding: 12px 16px;
+  height: 116px;
   border: 1px solid var(--gray-lighter);
   border-radius: 4px;
   overflow-y: auto;
