+++ src/app/myreservation/style.module.css
@@ -173,3 +173,13 @@
   box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.3);
   z-index: 10;
 }
+
+.profileCard {
+  display: flex;
+}
+
+@media (max-width: 430px) {
+  .profileCard {
+    display: none;
+  }
+}
