+++ src/components/ProfileCard/ProfileCard.module.css
@@ -54,8 +54,8 @@
   }
 }
 
-@media (max-width: 450px) {
-  .container {
-    display: none;
+@media (max-width: 430px) {
+  .card {
+    display: none !important;
   }
 }
